Changed metro_simulator.go, advancedConfig.go, and config.go so that we don't have to work with the GOPATH anymore
basepath is no longer the GOPATH but the program execution directory. It can be slightly different depending on how we run the project: either it can be "pfe-2018-network-journey-simulator" when launching with an IDE runtime, or it can be "pfe-2018-network-journey-simulator/src/main" when launching the program from the command line. The change was therefore to make it "pfe-2018-network-journey-simulator" in all cases.

As a result, nobody has to put their GOPATH on the project for it to work (an advantage for those who have several Go projects that are not located in the same folder and don't want to change the GOPATH every time).