Installation of two flatlaf libraries for a more pleasing visual experience and to be able to apply themes to the application.
Addition of a JMenuBar implemented with the MenuBar class to have an application bar with: file management, configuration management and application theme change.

Change of GUI structure with the addition of the ToolBar class
It contains elements from the old GUI (adding stations, lines, etc.) as well as new elements (deleting stations, lines, etc.). The actions from the old GUI are still there, but the new buttons are visual only for the moment. Nothing will happen if you press them.

To fix :
HUD panel (still the old one for now)
Dobule click on an area in light mode : destination in dark mode
Code structuration to discuss
Buttons sizes/space taken