[12] [DEV] [0.0.7]
Functionality for the distribution set has changed.
From now on, a total is displayed with the total of each distribution. if the total is equal to 100, it is displayed in green, otherwise in red.
When OK is clicked, if only one of the totals is not equal to 100, an error message is displayed for that total and the values are not saved, and the values for the other distribution are saved.
If both totals are not equal to 100, two error messages are displayed and no values are saved.