[38] [DEV] [0.0.4]
Now, when we export our network, we put in the xml the geographical position in the center of the panel as well as the current zoom.

This data is reused when importing a network to "teleport" to the position of the network in the world and at the appropriate zoom level.